requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.0
python-dotenv==1.0.0
httpx>=0.28.0,<1.0.0
pyjwt==2.8.0
cryptography>=43.0.1
okta-ai-sdk-proto==1.0.1
